--- a/modules/core/include/opencv2/core/cv_cpu_dispatch.h
+++ b/modules/core/include/opencv2/core/cv_cpu_dispatch.h
@@ -82,9 +82,9 @@
 #  include <immintrin.h>
 #  define CV_AVX2 1
 #endif
-#ifdef CV_CPU_COMPILE_AVX512
+#ifdef CV_CPU_COMPILE_AVX_512F
 #  include <immintrin.h>
-#  define CV_AVX512 1
+#  define CV_AVX_512F 1
 #endif
 #ifdef CV_CPU_COMPILE_FMA3
 #  define CV_FMA3 1