This pull request adds support for AVX512 instructions for some of the DNN operations

```
allow_multiple_commits=1
```