@alalek 2017-12-26
Thank you for the contribution!

May I ask you to share performance numbers for AVX2 vs AVX512.

Something like these:
- setup env: `OPENCV_TEST_DATA_PATH=<opencv_extra>/testdata`
- setup env: `OPENCV_DNN_TEST_DATA_PATH=<some_dir_with_dnn_subfolder>`
  use this script to download test DNN models (into "dnn" subfolder, size ~2Gb): https://github.com/opencv/opencv_extra/blob/master/testdata/dnn/download_models.py
- compile baseline code (without patch) and run:
    ```
    ./bin/opencv_perf_dnn --gtest_output=xml:base.xml
    ```
- compile patched code and run:
    ```
    ./bin/opencv_perf_dnn --gtest_output=xml:optimized.xml
    ```
- generate report:
    ```
    python <opencv_src>/modules/ts/misc/summary.py base.xml optimized.xml
    ```
- post results here (use "-o markdown" to generate report compatible with GitHub comments)